pretraitement/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.11
apscheduler==3.10.4
python-dotenv==1.2.1
